Add files via upload
This code processes ocean-colour data, which can be downloaded with the script 'Download_ocean_colour_data.R'.
A Principal Component Analysis is performed, and it is found that the leading source of variance in ocean-colour in the sample image is likely to be caused by enhanced scattering by calcifying phytoplankton in the surface ocean. 
This source of variance may bias the inference of commonly employed derived variables in biological oceanography, such as chlorophyll-a concentration. 
The Principal Component Analysis is used, thereafter, to define a threshold to mask pixels where ocean-colour is strongly affected by calcifying phytoplankton.