Add files via upload
This is an example script to illustrate how Redundancy Analysis is performed and its utilities.
In this example, we develop further by exploring the capacity for different Redundancy Analysis results to be compared.
We might have multiple different ways of describing the dependent variables, for example, and want to know
whether this changes the way the explanatory variables constrain variance. 
In this example we will explore whether geography constrains variance in sea surface temperature and storminess more similarly
in 2002 and 2003, than it does compared with 2018. 
We will also use various permutation based approaches to explore the variance explained by individual explanatory variables (latitude and longitude).
 We are going to use the dataset 'SST_and_storms.csv'
 This is a dataset I produced by downloading sea surface temperature data
 from the ARCTIC_REANALYSIS_PHYS_002_003 dataset, 
 and wind-speed from the WIND_GLO_WIND_L3_REP_OBSERVATIONS_012_005 dataset.
 Both of these datasets are available on https://marine.copernicus.eu/
 I averaged data for September, over a series of different years, 
 within 1 degree latitude-longitude boxes, so that the dataset is 
 small and manageable, and so that researchers interested in the original
 data will be motivated to visit https://marine.copernicus.eu/
 
 The csv file contains columns for latitude, longitude, year, 
 sea surface temperature, and storminess. 
 storminess is computed as the average number of days,
 within a grid cell, that wind speeds exceed 10 metres per second.